% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbcbd_utils.R
\name{gbcbd_get_default_cache_folder}
\alias{gbcbd_get_default_cache_folder}
\title{Returns default cache folder}
\usage{
gbcbd_get_default_cache_folder()
}
\value{
A path to directory
}
\description{
Internal use. USe the same default cache folder for all other functions.
}
\examples{
my.cache.folder <- gbcbd_get_default_cache_folder()
}
