% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbcbd_utils.R
\name{gbcbd_get_JSON_fct}
\alias{gbcbd_get_JSON_fct}
\title{Function for switching functions depending on using cache or not}
\usage{
gbcbd_get_JSON_fct(use.memoise = TRUE,
  cache.path = gbcbd_get_default_cache_folder())
}
\arguments{
\item{use.memoise}{Logical. Sets the use of caching system - TRUE (default) or FALSE}

\item{cache.path}{Path to save cache files - 'rbcb2_cache' (default)}
}
\value{
A JSON function that will use cache or not
}
\description{
Internal use. Simply switches a function given a choice for using memoise.
}
\examples{

fct_get_JSON <- gbcbd_get_JSON_fct()
}
