% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbcbd_utils.R
\name{gbcbd_message}
\alias{gbcbd_message}
\title{Custom function for printing messages}
\usage{
gbcbd_message(str.in, be.quiet)
}
\arguments{
\item{str.in}{Message string}

\item{be.quiet}{Logical. Should print it or not?}
}
\value{
Nothing
}
\description{
Custom function for printing messages
}
\examples{
gbcbd_message('Test', FALSE)
}
