% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbcbd_utils.R
\name{gbcbd_test_internet}
\alias{gbcbd_test_internet}
\title{Tests for an internet conection}
\usage{
gbcbd_test_internet()
}
\value{
Logical TRUE/FALSE
}
\description{
A helper function that testa and make sure the user has a live internet conection.
}
\examples{

inet.flag <- gbcbd_test_internet()
}
